perf(sql): more efficient hash table implementations
