+++ benchmarks/src/main/java/org/questdb/GroupByLong128HashSetBenchmark.java
@@ -27,7 +27,7 @@ package org.questdb;
 import io.questdb.cairo.ColumnType;
 import io.questdb.cairo.SingleColumnType;
 import io.questdb.cairo.map.MapKey;
-import io.questdb.cairo.map.OrderedMap;
+import io.questdb.cairo.map.VarSizeMap;
 import io.questdb.griffin.engine.groupby.GroupByAllocator;
 import io.questdb.griffin.engine.groupby.GroupByAllocatorArena;
 import io.questdb.griffin.engine.groupby.GroupByLong128HashSet;
@@ -49,7 +49,7 @@ public class GroupByLong128HashSetBenchmark {
     private static final double loadFactor = 0.7;
     private static final GroupByLong128HashSet groupByLong128HashSet = new GroupByLong128HashSet(64, loadFactor, 0);
     private static final int orderedMapPageSize = 1024 * 1024;
-    private static final OrderedMap orderedMap = new OrderedMap(orderedMapPageSize, new SingleColumnType(ColumnType.UUID), null, 64, loadFactor, Integer.MAX_VALUE);
+    private static final VarSizeMap orderedMap = new VarSizeMap(orderedMapPageSize, new SingleColumnType(ColumnType.UUID), null, 64, loadFactor, Integer.MAX_VALUE);
     private static final Rnd rnd = new Rnd();
     private static long ptr = 0;
     @Param({"2500", "25000", "250000", "2500000"})
