+++ core/src/main/java/io/questdb/PropertyKey.java
@@ -60,9 +60,9 @@ public enum PropertyKey implements ConfigPropertyKey {
     CAIRO_CHARACTER_STORE_CAPACITY("cairo.character.store.capacity"),
     CAIRO_CHARACTER_STORE_SEQUENCE_POOL_CAPACITY("cairo.character.store.sequence.pool.capacity"),
     CAIRO_COLUMN_POOL_CAPACITY("cairo.column.pool.capacity"),
-    CAIRO_COMPACT_MAP_LOAD_FACTOR("cairo.compact.map.load.factor"),
     CAIRO_EXPRESSION_POOL_CAPACITY("cairo.expression.pool.capacity"),
-    CAIRO_FAST_MAP_LOAD_FACTOR("cairo.fast.map.load.factor"),
+    CAIRO_FIXED_SIZE_MAP_LOAD_FACTOR("cairo.fixed.size.map.load.factor"),
+    CAIRO_VAR_SIZE_MAP_LOAD_FACTOR("cairo.var.size.map.load.factor"),
     CAIRO_SQL_JOIN_CONTEXT_POOL_CAPACITY("cairo.sql.join.context.pool.capacity"),
     CAIRO_LEXER_POOL_CAPACITY("cairo.lexer.pool.capacity"),
     CAIRO_SQL_MAP_KEY_CAPACITY("cairo.sql.map.key.capacity"),
