+++ core/src/main/java/io/questdb/griffin/engine/groupby/DistinctTimeSeriesRecordCursorFactory.java
@@ -27,7 +27,7 @@ package io.questdb.griffin.engine.groupby;
 import io.questdb.cairo.*;
 import io.questdb.cairo.map.Map;
 import io.questdb.cairo.map.MapKey;
-import io.questdb.cairo.map.OrderedMap;
+import io.questdb.cairo.map.VarSizeMap;
 import io.questdb.cairo.sql.Record;
 import io.questdb.cairo.sql.*;
 import io.questdb.griffin.PlanSink;
@@ -56,7 +56,7 @@ public class DistinctTimeSeriesRecordCursorFactory extends AbstractRecordCursorF
             // sink will be storing record columns to map key
             columnFilter.of(metadata.getColumnCount());
             RecordSink recordSink = RecordSinkFactory.getInstance(asm, metadata, columnFilter, false);
-            Map dataMap = new OrderedMap(
+            Map dataMap = new VarSizeMap(
                     configuration.getSqlSmallMapPageSize(),
                     metadata,
                     configuration.getSqlDistinctTimestampKeyCapacity(),
