+++ core/src/main/java/io/questdb/griffin/engine/join/JoinRecordMetadata.java
@@ -28,7 +28,7 @@ import io.questdb.cairo.*;
 import io.questdb.cairo.map.Map;
 import io.questdb.cairo.map.MapKey;
 import io.questdb.cairo.map.MapValue;
-import io.questdb.cairo.map.OrderedMap;
+import io.questdb.cairo.map.VarSizeMap;
 import io.questdb.cairo.sql.RecordMetadata;
 import io.questdb.std.Chars;
 import io.questdb.std.MemoryTag;
@@ -45,7 +45,7 @@ public class JoinRecordMetadata extends AbstractRecordMetadata implements Closea
     private int refCount;
 
     public JoinRecordMetadata(CairoConfiguration configuration, int columnCount) {
-        this.map = new OrderedMap(
+        this.map = new VarSizeMap(
                 configuration.getSqlJoinMetadataPageSize(),
                 keyTypes,
                 valueTypes,
