+++ core/src/main/resources/io/questdb/site/conf/server.conf
@@ -227,8 +227,11 @@ query.timeout.sec=60
 # size of the ExpressionNode pool in SqlCompiler
 #cairo.expression.pool.capacity=8192
 
-# load factor for all FastMaps
-#cairo.fast.map.load.factor=0.7
+# load factor for fixed-size key Maps
+#cairo.var.size.map.load.factor=0.6
+
+# load factor for var-size key Maps
+#cairo.var.size.map.load.factor=0.7
 
 # size of the JoinContext pool in SqlCompiler
 #cairo.sql.join.context.pool.capacity=64
