+++ core/src/test/resources/server.conf
@@ -76,7 +76,8 @@ cairo.character.store.capacity=2048
 cairo.character.store.sequence.pool.capacity=128
 cairo.column.pool.capacity=2048
 cairo.expression.pool.capacity=1024
-cairo.fast.map.load.factor=0.3
+cairo.fixed.size.map.load.factor=0.3
+cairo.var.size.map.load.factor=0.4
 cairo.sql.join.context.pool.capacity=32
 cairo.lexer.pool.capacity=1024
 cairo.sql.map.key.capacity=1024
